chore: Fixed int viper conversion error

+++ pkg/porter/service.go
@@ -1,7 +1,7 @@
 package porter
 
 type ServiceOptions struct {
-	Port        int32
+	Port        int64
 	ServiceName string
 }
